# clean-architecture-sample/docs/swagger.yaml
definitions:
  api.CreateUserRequest:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  api.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
  api.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  handler.CreateUserRequest:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  handler.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
  handler.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
  license:
    name: yyamakura
  title: gin-swagger sample
  version: "1.0"
paths:
  /users:
    post:
      consumes:
      - application/json
      description: 新しいユーザーを登録します
      parameters:
      - description: ユーザー情報
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ユーザー作成
      tags:
      - users
  /users/{id}:
    delete:
      description: 指定IDのユーザーを削除します
      parameters:
      - description: ユーザーID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ユーザー削除
      tags:
      - users
    get:
      description: 指定IDのユーザー情報を取得します
      parameters:
      - description: ユーザーID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ユーザー取得
      tags:
      - users
    put:
      consumes:
      - application/json
      description: 指定IDのユーザー情報を更新します
      parameters:
      - description: ユーザーID
        in: path
        name: id
        required: true
        type: string
      - description: ユーザー情報
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ユーザー更新
      tags:
      - users
swagger: "2.0"